requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.8
python-multipart==0.0.6
passlib[argon2]==1.7.4
python-jose==3.3.0
pydantic==2.7.0
python-dotenv==1.0.0
pydantic-settings==2.13.1
authlib==1.2.1
httpx==0.25.0
